Part 4 complete

+++ part4/controllers/blogs.js
@@ -1,37 +1,55 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
-blogsRouter.get('/', (request, response) => {
-  Blog
-    .find({})
-    .then(blogs => {
-      response.json(blogs)
-    })
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user', { passwordHash: 0, blogs: 0 } )
+  response.json(blogs.map(blog => blog.toJSON()))
 })
 
-blogsRouter.post('/', (request, response) => {
-  let blog = {}
-  if (!request.body.title && !request.body.url) {
-    return response.status(400).end()
-  }
-  if (!request.body.likes) {
-    blog = new Blog({ ...request.body, likes: 0 })
-  }
-  else {
-    blog = new Blog(request.body)
-  }
+blogsRouter.post('/', async (request, response) => {
+  const body = request.body
 
-  blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
-    })
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    let blog = {}
+    let user = {}
+    if (!body.title && !body.url) {
+      return response.status(400).end()
+    }
+    if (!body.likes) {
+      blog = new Blog({ ...request.body, likes: 0 })
+    }
+    else {
+      user = await User.findOne({ username: 'root' })
+      blog = new Blog({ ...request.body, user: user._id })
+    }
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.json(savedBlog.toJSON())
+  } catch (exception) {
+    return response.status(404).json(exception)
+  }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if ( blog.user.toString() === decodedToken.id.toString() ) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
+    response.status(403).json({ error: 'blog can only be deleted by its own user' })
   }
   catch (err) {
     return err
@@ -50,8 +68,8 @@ blogsRouter.put('/:id', async (request, response) => {
     const res = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
     response.json(res.toJSON())
   }
-  catch (err) {
-    return err
+  catch (exception) {
+    return exception
   }
 })
